+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/common/exception/TestCertificateServerException.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.covidcertificate.exception
+package de.rki.coronawarnapp.covidcertificate.common.exception
 
 import android.content.Context
 import androidx.annotation.StringRes
